Don't use int_fast8_t when checking if a mouse button is valid
Fixes buffer overflow crash in Super Chicken Jumper

+++ src/runner_mouse.c
@@ -6,11 +6,11 @@
 #include <stdlib.h>
 #include <string.h>
 
-static bool isValidButtonVirtual(int_fast8_t button) {
+static bool isValidButtonVirtual(int32_t button) {
     return button >= -1 && GML_MOUSE_BUTTONS > button;
 }
 
-static bool isValidButton(int_fast8_t button) {
+static bool isValidButton(int32_t button) {
     return button >= 1 && GML_MOUSE_BUTTONS > button;
 }
